chore: report, restriction 분리 및 코드 수정
