+++ src/main/java/eureca/capstone/project/admin/report/repository/ReportTypeRepository.java
@@ -4,4 +4,5 @@ import eureca.capstone.project.admin.report.entity.ReportType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportTypeRepository extends JpaRepository<ReportType, Long> {
+    ReportType findByType(String name);
 }
