+++ src/main/java/eureca/capstone/project/admin/report/service/ReportService.java
@@ -12,11 +12,7 @@ public interface ReportService {
 
     ReportCountDto getReportCounts();
     Page<ReportHistoryDto> getReportHistoryListByStatusCode(String statusCode, String keyword, Pageable pageable);
-    Page<RestrictionDto> getRestrictionListByStatusCode(String statusCode,String keyword, Pageable pageable);
     void processReportByAdmin(Long reportHistoryId, ProcessReportDto request);
     void createReportAndProcessWithAI(Long userId, Long transactionFeedId, Long reportTypeId, String reason);
-    void acceptRestrictions(Long restrictionTargetId);
-    void rejectRestrictions(Long restrictionTargetId);
     ReportDetailResponseDto getReportDetail(Long reportId);
-    List<RestrictionReportResponseDto> getRestrictionReportHistory(Long restrictionId);
 }
